// core/types/generated/contentful.d.ts
export interface ICtaFields {
  /** Text */
  text: string;

  /** Internal Reference */
  internalReference?:
    | IProductFinishedGoods
    | IProductPartsAndAccessories
    | IPageStandard
    | undefined;

  /** External Link */
  externalLink?: string | undefined;
}

export interface ICta extends Entry<ICtaFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "cta";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IHeroCarouselFields {
  /** Hero Slides */
  heroSlides?: IHeroSlide[] | undefined;
}

export interface IHeroCarousel extends Entry<IHeroCarouselFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "heroCarousel";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IHeroSlideFields {
  /** Image */
  image?: Asset | undefined;

  /** Headline */
  headline: string;

  /** Subhead */
  subhead?: string | undefined;

  /** CTA Label */
  ctaLabel?: string | undefined;

  /** CTA Link */
  ctaLink?: Entry<{ [fieldId: string]: unknown }> | undefined;
}

export interface IHeroSlide extends Entry<IHeroSlideFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "heroSlide";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IInspirationBentoFields {
  /** Heading */
  heading?: string | undefined;

  /** CTA */
  cta?: ICta | undefined;

  /** Video */
  video?: string | undefined;

  /** Inspiration Cards */
  inspirationCards?: IInspirationCard[] | undefined;
}

export interface IInspirationBento extends Entry<IInspirationBentoFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "inspirationBento";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IInspirationCardFields {
  /** Title */
  title: string;

  /** Content Reference */
  contentReference: ITutorial | IRecipe;
}

export interface IInspirationCard extends Entry<IInspirationCardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "inspirationCard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPageStandardFields {
  /** Page Name */
  pageName: string;

  /** Meta Title (SEO) */
  metaTitleSeo: string;

  /** Meta Description (SEO) */
  metaDescription?: string | undefined;

  /** Meta Keywords (SEO) */
  metaKeywordsSeo?: string | undefined;

  /** Page Slug */
  pageSlug: string;

  /** Optional Page Description */
  optionalPageDescription?: Document | undefined;

  /** Page Content Entries */
  pageContent?: Entry<{ [fieldId: string]: unknown }>[] | undefined;
}

export interface IPageStandard extends Entry<IPageStandardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pageStandard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}